+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.0.23/1.7.23/1.6.23/1.5.23 04.2026
 - Entitiy debug route
+- Storefront Twig files debug route
+- Removed deprecated ReqserWebhookService
 
 ## 2.0.22/1.7.22/1.6.22/1.5.22 04.2026
 - Remove language switcher flag override logic